pyproject.toml
[tool.poetry]
name = "GAA Computer Vision Model"
version = "0.1.0"
description = ""
authors = ["Donal Conlon - Sportedge, donal@sportedge.io"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12.0"
norfair = {git = "https://github.com/tryolabs/norfair.git", rev = "master"}
torch = "^1.12.1"
opencv-python = "^4.6.0.66"
numpy = "^1.23.4"
Pillow = "^9.2.0"
matplotlib = "^3.6.1"
pandas = "^1.5.0"
torchvision = "^0.13.1"
PyYAML = "^6.0"
tqdm = "^4.64.1"
seaborn = "^0.12.0"
ipython = "^8.5.0"
sahi = "^0.11.14"
supervision = "^0.14.0"
psutil = "^5.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
